refactor resultset and formatter classes to functional modules
