--- nerium/contrib/formatter/__init__.py
@@ -1,5 +1,0 @@
-from nerium.contrib.formatter.affix import AffixFormatter
-from nerium.contrib.formatter.compact import CompactFormatter
-from nerium.contrib.formatter.csv import CsvFormatter
-from nerium.contrib.formatter.default import DefaultFormatter
-from nerium.contrib.formatter.sum import SumFormatter
