--- nerium/contrib/formatter/default.py
@@ -1,8 +1,0 @@
-from nerium import ResultFormatter
-
-
-class DefaultFormatter(ResultFormatter):
-    """ No-op default
-    """
-    def format_results(self):
-        return self.result
