--- nerium/contrib/resultset/__init__.py
@@ -1,2 +1,0 @@
-from nerium.contrib.resultset.sql import SQLResultSet
-from nerium.contrib.resultset.jinja import JinjaSQLResultSet
