+++ nerium/formatter/default.py
@@ -1,0 +1,6 @@
+""" No-op default
+"""
+
+
+def format_results(result, **kwargs):
+    return result
