+++ nerium/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.2b1'
+__version__ = '0.3'
